+++ healvis/tests/test_simulator.py
@@ -9,13 +9,13 @@ import healpy as hp
 import os
 import yaml
 import shutil
-import nose.tools as nt
 from astropy.cosmology import Planck15
 
 from pyuvdata import UVData
 
 from healvis import sky_model, simulator, observatory, beam_model, utils
 from healvis.data import DATA_PATH
+import healvis.tests as simtest
 
 
 def test_setup_uvdata():
@@ -24,7 +24,7 @@ def test_setup_uvdata():
                                  telescope_location=(-30.72152777777791, 21.428305555555557, 1073.0000000093132),
                                  telescope_name="HERA", Nfreqs=10, start_freq=1e8, bandwidth=1e8, Ntimes=60,
                                  time_cadence=100.0, start_time=2458101.0, pols=['xx'], no_autos=True, make_full=True, run_check=True)
-    nt.assert_equal(uvd.Nbls, uvd.Nants_data * (uvd.Nants_data - 1) / 2)
+    assert uvd.Nbls == uvd.Nants_data * (uvd.Nants_data - 1) / 2
 
     # check selection works
     bls = [(0, 11), (0, 12), (0, 13)]
@@ -32,13 +32,13 @@ def test_setup_uvdata():
                                  telescope_location=(-30.72152777777791, 21.428305555555557, 1073.0000000093132),
                                  telescope_name="HERA", Nfreqs=10, start_freq=1e8, bandwidth=1e8, Ntimes=60,
                                  time_cadence=100.0, start_time=2458101.0, pols=['xx'], bls=bls, make_full=True, run_check=True)
-    nt.assert_equal(uvd.Nbls, len(bls))
+    assert uvd.Nbls == len(bls)
     uvd = simulator.setup_uvdata(array_layout=os.path.join(DATA_PATH, "configs/HERA65_layout.csv"),
                                  telescope_location=(-30.72152777777791, 21.428305555555557, 1073.0000000093132),
                                  telescope_name="HERA", Nfreqs=10, start_freq=1e8, bandwidth=1e8, Ntimes=60,
                                  time_cadence=100.0, start_time=2458101.0, pols=['xx'], bls=bls, make_full=True, antenna_nums=[11],
                                  no_autos=False, run_check=True)
-    nt.assert_equal(uvd.Nbls, 1)
+    assert uvd.Nbls == 1
 
 
 def test_run_simulation():
@@ -63,15 +63,15 @@ def test_run_simulation():
     uvd.read(os.path.join(param_dict['filing']['outdir'], "test_sim.uvh5"))
 
     # basic checks
-    nt.assert_equal(uvd.Nfreqs, 10)
-    nt.assert_equal(uvd.Ntimes, 5)
-    nt.assert_equal(uvd.Nbls, 4)
-    nt.assert_equal(uvd.Npols, 2)
-    nt.assert_true('foo' in uvd.history)  # check add_to_history was propagated
-    nt.assert_true("SKYPARAM" in uvd.history)  # check param_dict was written to history
+    assert uvd.Nfreqs == 10
+    assert uvd.Ntimes == 5
+    assert uvd.Nbls == 4
+    assert uvd.Npols == 2
+    assert 'foo' in uvd.history  # check add_to_history was propagated
+    assert "SKYPARAM" in uvd.history  # check param_dict was written to history
 
     # test data_array ordering is correct--i.e. Nbls, Ntimes--by asserting that auto-correlation is purely real for all times
-    nt.assert_true(np.isclose(uvd.get_data(0, 0).imag, 0.0).all())
+    assert np.isclose(uvd.get_data(0, 0).imag, 0.0).all()
 
     # erase output directory
     shutil.rmtree(param_dict['filing']['outdir'])
@@ -89,7 +89,7 @@ def test_run_simulation_partial_freq():
                                  telescope_location=(-30.72152777777791, 21.428305555555557, 1073.0000000093132),
                                  telescope_name="HERA", Ntimes=60, time_cadence=100.0, start_time=2458101.0,
                                  pols=['xx'], bls=bls, make_full=True, run_check=True, freq_array=sky.freqs)
-    test_uvh5 = os.path.join(DATA_PATH, "test_freq_parallel_sim.uvh5")
+    test_uvh5 = os.path.join(simtest.TESTDATA_PATH, "test_freq_parallel_sim.uvh5")
     uvd.write_uvh5(test_uvh5, clobber=True)
 
     # run simulation
@@ -102,8 +102,8 @@ def test_run_simulation_partial_freq():
     uvd2 = UVData()
     uvd2.read_uvh5(test_uvh5)
     d = uvd2.get_data(0, 11)
-    nt.assert_false(np.isclose(d[:, freq_chans], 0.0).all())
-    nt.assert_true(np.isclose(d[:, freq_chans[-1] + 1:], 0.0).all())
+    assert not (np.isclose(d[:, freq_chans], 0.0).all())
+    assert np.isclose(d[:, freq_chans[-1] + 1:], 0.0).all()
 
     # clean up
     os.remove(test_uvh5)
@@ -133,14 +133,14 @@ def test_setup_light_uvdata():
 
     obs_full = simulator.setup_observatory_from_uvdata(uvd_full, fov=30, set_pointings=True, beam=beam_type)
 
-    nt.assert_true(len(uvd.baseline_array) == len(bls))
-    nt.assert_true(len(uvd.time_array) == uvd.Ntimes)       # If the "light" weren't enabled, these would be length Nblts
+    assert len(uvd.baseline_array) == len(bls)
+    assert len(uvd.time_array) == uvd.Ntimes      # If the "light" weren't enabled, these would be length Nblts
 
     # Check that baselines, times, and frequencies all match
     for bi, bl in enumerate(bls):
-        nt.assert_true(np.all(obs_full.array[bi].enu == obs.array[bi].enu))
+        assert np.all(obs_full.array[bi].enu == obs.array[bi].enu)
 
-    nt.assert_true(np.all(obs_full.freqs == obs.freqs))
+    assert np.all(obs_full.freqs == obs.freqs)
 
 
 def test_parse_freq_params():
@@ -157,33 +157,36 @@ def test_parse_freq_params():
     # case i
     fdict = simulator.parse_frequency_params(dict(start_freq=start_freq, Nfreqs=Nfreqs, channel_width=channel_width))
     for k in master_fdict:
-        nt.assert_true(np.all(np.isclose(master_fdict[k], fdict[k])))
+        assert np.all(np.isclose(master_fdict[k], fdict[k]))
     # case ii
     fdict = simulator.parse_frequency_params(dict(start_freq=start_freq, Nfreqs=Nfreqs, bandwidth=bandwidth))
     for k in master_fdict:
-        nt.assert_true(np.all(np.isclose(master_fdict[k], fdict[k])))
+        assert np.all(np.isclose(master_fdict[k], fdict[k]))
     # case iii
     fdict = simulator.parse_frequency_params(dict(start_freq=start_freq, Nfreqs=Nfreqs, end_freq=end_freq))
     for k in master_fdict:
-        nt.assert_true(np.all(np.isclose(master_fdict[k], fdict[k])))
+        assert np.all(np.isclose(master_fdict[k], fdict[k]))
     # case iv
     fdict = simulator.parse_frequency_params(dict(start_freq=start_freq, end_freq=end_freq, channel_width=channel_width))
     for k in master_fdict:
-        nt.assert_true(np.all(np.isclose(master_fdict[k], fdict[k])))
+        assert np.all(np.isclose(master_fdict[k], fdict[k]))
 
     # test that if freq_array is present, it supercedes!
     _farray = np.linspace(100, 200, 10, endpoint=False)
     fdict = simulator.parse_frequency_params(dict(start_freq=start_freq, end_freq=end_freq, channel_width=channel_width, freq_array=_farray))
-    nt.assert_true(np.all(np.isclose(_farray, fdict['freq_array'])))
+    assert np.all(np.isclose(_farray, fdict['freq_array']))
 
     # test Nfreqs = 1 exception
-    nt.assert_raises(ValueError, simulator.parse_frequency_params, dict(freq_array=np.array([100.0])))
+    simtest.assert_raises_message(ValueError, 'Channel width must be specified if passed freq_arr has length 1',
+                                  simulator.parse_frequency_params, dict(freq_array=np.array([100.0])))
 
     # test evenly divisible exception
-    nt.assert_raises(ValueError, simulator.parse_frequency_params, dict(start_freq=100, end_freq=101, channel_width=0.33))
+    simtest.assert_raises_message(ValueError, 'end_freq - start_freq must be evenly divisible by channel_width',
+                                  simulator.parse_frequency_params, dict(start_freq=100, end_freq=101, channel_width=0.33))
 
     # test improper combination KeyError
-    nt.assert_raises(KeyError, simulator.parse_frequency_params, dict(start_freq=100.0))
+    simtest.assert_raises_message(KeyError, "Couldn't find any proper combination of keys in freq_params",
+                                  simulator.parse_frequency_params, dict(start_freq=100.0))
 
 
 def test_parse_time_params():
@@ -200,33 +203,36 @@ def test_parse_time_params():
     # case i
     tdict = simulator.parse_time_params(dict(start_time=start_time, Ntimes=Ntimes, time_cadence=time_cadence))
     for k in master_tdict:
-        nt.assert_true(np.all(np.isclose(master_tdict[k], tdict[k])))
+        assert np.all(np.isclose(master_tdict[k], tdict[k]))
     # case ii
     tdict = simulator.parse_time_params(dict(start_time=start_time, Ntimes=Ntimes, duration=duration))
     for k in master_tdict:
-        nt.assert_true(np.all(np.isclose(master_tdict[k], tdict[k])))
+        assert np.all(np.isclose(master_tdict[k], tdict[k]))
     # case iii
     tdict = simulator.parse_time_params(dict(start_time=start_time, Ntimes=Ntimes, end_time=end_time))
     for k in master_tdict:
-        nt.assert_true(np.all(np.isclose(master_tdict[k], tdict[k])))
+        assert np.all(np.isclose(master_tdict[k], tdict[k]))
     # case iv
     tdict = simulator.parse_time_params(dict(start_time=start_time, end_time=end_time, time_cadence=time_cadence))
     for k in master_tdict:
-        nt.assert_true(np.all(np.isclose(master_tdict[k], tdict[k])))
+        assert np.all(np.isclose(master_tdict[k], tdict[k]))
 
     # test that if time_array is present, it supercedes!
     _tarray = np.linspace(2458201, 2458201.10, 10, endpoint=False)
     tdict = simulator.parse_time_params(dict(start_time=start_time, end_time=end_time, time_cadence=time_cadence, time_array=_tarray))
-    nt.assert_true(np.all(np.isclose(_tarray, tdict['time_array'])))
+    assert np.all(np.isclose(_tarray, tdict['time_array']))
 
     # test Ntimes = 1 exception
-    nt.assert_raises(ValueError, simulator.parse_time_params, dict(time_array=np.array([100.0])))
+    simtest.assert_raises_message(ValueError, 'time_cadence must be specified if Ntimes == 1',
+                                  simulator.parse_time_params, dict(time_array=np.array([100.0])))
 
     # test evenly divisible exception
-    nt.assert_raises(ValueError, simulator.parse_time_params, dict(start_time=100, end_time=101, time_cadence=0.33))
+    simtest.assert_raises_message(ValueError, 'end_time - start_time must be evenly divisible by time_cadence',
+                                  simulator.parse_time_params, dict(start_time=100, end_time=101, time_cadence=0.33))
 
     # test improper combination KeyError
-    nt.assert_raises(KeyError, simulator.parse_time_params, dict(start_time=100.0))
+    simtest.assert_raises_message(KeyError, "Couldn't find any proper combination of keys in time_params.",
+                                  simulator.parse_time_params, dict(start_time=100.0))
 
 
 def test_redundant_setup():
@@ -258,8 +264,8 @@ def test_redundant_setup():
                                   freq_array=freq_array, time_array=time_array, redundancy=redtol, no_autos=False)
 
     uvd0 = simulator.complete_uvdata(uvd0)
-    nt.assert_true(uvd0.Nbls == 62)
-    nt.assert_true(uvd.Nbls == 61)
+    assert uvd0.Nbls == 62
+    assert uvd.Nbls == 61
 
 
 def test_freq_time_params():
@@ -271,5 +277,5 @@ def test_freq_time_params():
     freq_dict = utils.freq_array_to_params(freqs)
     ftest = simulator.parse_frequency_params(freq_dict)
     ttest = simulator.parse_time_params(time_dict)
-    nt.assert_true(np.allclose(ftest['freq_array'], freqs))
-    nt.assert_true(np.allclose(ttest['time_array'], times))
+    assert np.allclose(ftest['freq_array'], freqs)
+    assert np.allclose(ttest['time_array'], times)