+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Option to let pixels smoothly `set` with `horizon taper`
 
 ### Changed
+- Use pytest instead of nosetests
 - UVBeam objects now have data arrays moved to shared memory.
 
 ### Fixed